Adding support for schema via opa eval -s
